style:распределить пакеты и файлы

+++ main-srvc/src/main/java/ru/practicum/part/admin_part/compilations/service/AdminCompilationsServiceImpl.java
@@ -3,11 +3,13 @@ package ru.practicum.part.admin_part.compilations.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.dto.compilation.CompilationDto;
+import ru.practicum.dto.compilation.CompilationMapper;
 import ru.practicum.dto.compilation.CompilationOutDto;
 import ru.practicum.dto.event.EventMapper;
 import ru.practicum.dto.event.ShortEventDto;
 import ru.practicum.exception.BadRequestException;
 import ru.practicum.exception.ConflictException;
+import ru.practicum.exception.NotFoundException;
 import ru.practicum.model.compilation.Compilation;
 import ru.practicum.model.event.Event;
 import ru.practicum.storage.compilation.CompilationRepository;
@@ -18,7 +20,7 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class CompilationsServiceImpl implements CompilationsService {
+public class AdminCompilationsServiceImpl implements AdminCompilationsService {
     private final EventRepository eventRepository;
     private final CompilationRepository compilationRepository;
 
@@ -33,9 +35,7 @@ public class CompilationsServiceImpl implements CompilationsService {
                 || compilationDto.getTitle().isEmpty()) {
             throw new BadRequestException("Field title filled incorrectly");
         }
-        if (compilationDto.getTitle().length() > 50) {
-            throw new BadRequestException("Title is too long");
-        }
+        compNameValid(compilationDto);
 
         Compilation compilation = new Compilation();
         if (compilationDto.getPinned() != null) {
@@ -47,59 +47,49 @@ public class CompilationsServiceImpl implements CompilationsService {
         compilationRepository.save(compilation);
 
         List<ShortEventDto> events = new ArrayList<>();
-
         if (compilationDto.getEvents() != null) {
             for (Integer eventId : compilationDto.getEvents()) {
                 Event event = eventRepository.findById(eventId).get();
                 events.add(EventMapper.toEventShortDto(event));
-
                 event.setCompilation(compilationRepository.findFirstByOrderByIdDesc());
                 eventRepository.save(event);
-
-                //eventRepository.findById(eventId).get().setCompilation(compilationRepository.findFirstByOrderByIdDesc());
             }
         }
 
-        CompilationOutDto compilationOutDto = new CompilationOutDto();
+        CompilationOutDto compilationOutDto = CompilationMapper.toCompOutDto(compilation);
         compilationOutDto.setEvents(events);
         compilationOutDto.setId(compilationRepository.findFirstByOrderByIdDesc().getId());
-        compilationOutDto.setPinned(compilation.getPinned());
-        compilationOutDto.setTitle(compilation.getTitle());
 
         return compilationOutDto;
     }
 
     @Override
     public void deleteComp(Integer compId) throws Exception {
-        try {
-            compilationRepository.deleteById(compId);
-        } catch (Exception e) {
-            throw new BadRequestException("Fields filled incorrectly");
+        if (compilationRepository.findById(compId).isEmpty()) {
+            throw new NotFoundException("Compilation with id=" + compId + " not found");
         }
+        compilationRepository.deleteById(compId);
     }
 
     @Override
     public CompilationOutDto editComp(Integer compId, CompilationDto compilationDto) throws Exception {
-
-
-        Compilation compilation = compilationRepository.findById(compId).get();
-
-        if (compilationDto.getTitle() != null) {
-            if (compilationDto.getTitle().length() > 50) {
-                throw new BadRequestException("Title is too long");
-            }
-            compilation.setTitle(compilationDto.getTitle());
+        if (compilationRepository.findById(compId).isEmpty()) {
+            throw new NotFoundException("Compilation with id=" + compId + " not found");
         }
 
+        Compilation compilation = compilationRepository.findById(compId).get();
+        if (compilationDto.getTitle() != null) {
+            compNameValid(compilationDto);
+            compilation.setTitle(compilationDto.getTitle());
+        }
         compilation.setPinned(compilationDto.getPinned());
         compilationRepository.save(compilation);
 
         List<ShortEventDto> events = new ArrayList<>();
-
         if (compilationDto.getEvents() != null) {
             for (Integer eventId : compilationDto.getEvents()) {
-                events.add(EventMapper.toEventShortDto(eventRepository.findById(eventId).get()));
-                Event event =  eventRepository.findById(eventId).get();
+                Event event = eventRepository.findById(eventId).get();
+                events.add(EventMapper.toEventShortDto(event));
                 event.setCompilation(compilationRepository.findById(compId).get());
                 eventRepository.save(event);
             }
@@ -109,14 +99,16 @@ public class CompilationsServiceImpl implements CompilationsService {
             }
         }
 
-
-        CompilationOutDto compilationOutDto = new CompilationOutDto();
+        CompilationOutDto compilationOutDto = CompilationMapper.toCompOutDto(compilation);
         compilationOutDto.setEvents(events);
         compilationOutDto.setId(compId);
-        compilationOutDto.setPinned(compilation.getPinned());
-        compilationOutDto.setTitle(compilation.getTitle());
 
         return compilationOutDto;
+    }
 
+    private void compNameValid(CompilationDto compDto) throws Exception {
+        if (compDto.getTitle().length() > 50) {
+            throw new BadRequestException("Title is too long");
+        }
     }
 }
